feat: add tabbed layout to instance page (closes #478)
Refactor the instance detail page into a tabbed layout with three tabs:
- Logs (default): expanded log viewer with more vertical space
- Trigger: full trigger details for the instance
- Telemetry: run info and token usage cards

Each tab has an inline SVG icon (document, bolt, chart-bar).
Follows the same pattern as AgentLayout with nested React Router routes.
Instance detail data is shared via InstanceContext to avoid re-fetching.
