+++ .changeset/instance-page-tabs.md
@@ -1,0 +1,5 @@
+---
+"@action-llama/action-llama": patch
+---
+
+Refactor the instance detail page into a tabbed layout (Logs, Trigger, Telemetry) with icons, following the same pattern as the agent page. The Logs tab is the default and gives the log viewer more vertical space. Closes #478.
